Potential fix for code scanning alert no. 3: Comparison of narrow type with wide type in loop condition

+++ web/src/main/java/ataf/web/pages/BasePage.java
@@ -692,7 +692,7 @@ public class BasePage {
      */
     public void waitForPageToLoad() {
         long millis = DEFAULT_IMPLICIT_WAIT_TIME;
-        int iterations = 1;
+        long iterations = 1L;
         do {
             try {
                 Thread.sleep(millis);
